EDSC-4638: Addressing feedback
